+++ S-et-L/src/connectors/TVConnector.java
@@ -32,7 +32,7 @@ implements TVI{
 	}
 
 	@Override
-	public int getCons() throws Exception {
+	public double getCons() throws Exception {
 		return ((TVI)this.offering).getCons();
 	}
